pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "hipaa-llm-wiki"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "ollama>=0.2.0",
    "pyyaml>=6.0",
    "click>=8.1",
    "rich>=13.0",
    "pdfplumber>=0.11.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[project.scripts]
wiki = "src.cli:cli"
